Added library functionality and fixed issues.
Improved the README.md for easier install.
Added .exe.manifest for the installer.exe [NOT INSTALLER OF THE APP]

Added library.html and its CSS
Created liraryprocess.js for better readability and impact on the library.
Fixed path issue.
Fixed naming issue, to go with electron naming of handlers.
Added creation of necessary files for info and downloaded games into the scraping file.